+++ extension/components/settings/index.ts
@@ -1,9 +1,8 @@
 // Settings components barrel export
 export * from './types'
 export { ProfilesTab } from './ProfilesTab'
-export { McpToolsTab } from './McpToolsTab'
 export { AudioTab } from './AudioTab'
 export { ImportExportDialog } from './ImportExportDialog'
 export { SettingsProvider, useSettings } from './SettingsContext'
 export type { SettingsContextValue } from './SettingsContext'
-export { ModalHeader, TokenHelpPanel, TabNavigation, ModalFooter } from './ModalUI'
+export { ModalHeader, TokenHelpPanel, ModalFooter } from './ModalUI'